Add prompt method
This method has two optional arguments: min and max. Min is set to 0
and max is set to 7; these values are used in a statement where asks
the player for a number between these two numbers, because the game
will ask two numbers for the coordinates of the grid in which to place
a token.