Made changes to Board class
First, added an attr_accessor to the grid instance variable.

The initialize method now only has the create_board method.

The create_board method now creates a 2d array of empty string values.

Added a new cell_empty? method that has two parameters: row and column,
values that are used to access a cell in the grid. The method then
returns true if this cell is an empty string, false otherwise.