Add verify_column method
This method receives a column index, then it obtains the column from
that index and finally checks if the column is not full. If it's not,
it returns the index. This is done to ensure that the player wants to
place a token on a column that is not full.