Add search_empty_cell method
This method receives a column index, and then it enters a loop from
the last row (5) to the first one, and searches with the cell_empty?
method the first empty cell of the column. When it founds it, it
returns a list with the coordinates of this cell to use it in the
put_token method.