Add player_turn method
In this method, it receives a player object, then it performs the
actions that make the player place a token on the board. In this
representation, it asks for a number, this number is the column
where the player wants to place his token.

It checks then if it's a valid input and if the column has available
cells; when it verifies that, it searches through a loop the lowest
available cell and finally it places the player's token there.