Add horizontal_line? method
This method checks if a row has a line of 4 consecutive tokens placed
by one player.

It has two parameters: row_index, the row that will check and token,
the token of the player that we are verifying if made a line.

First it gets the row with the obtain_row method, then it uses a loop
from 0 to 3 to get from the row four values (e.g. 1 through 4) and
then it checks if all the elements of this sub-array are the same. if
it founds the first group that met this condition, it returns true.

When the loop ends, we assume that the row doesn't have a line made, so
we return false.