Add diagonal_line? method
This method receives two parameters: diagonals, which are the list
or lists of diagonals values obtained from the grid. The other
parameter is token, the token that the function uses to check if
the player made a line.

First, it iterates through each diagonal list, then it obtains the
list length and then it checks if its 4, if it's then it only makes
a check with the all method if all are the same and returns true
when that condition is met. After this it skips to the next diagonal.

When the diagonal length is 5 or 6, it checks which one it is and it
changes the value to use it in another loop. If it's 5, it changes to
1 and if it's 6 it changes to 2.

Now it enters another loop, it takes from each iteration 4 values
and then checks with the all? method if all are the same. It returns
true when it founds the first group that met this condition.

When it finishes this loop, it goes to the next diagonal and repeats
all the procedure again.

If the method ends the diagonals loop, it means that the player didn't
fill a line, so it returns false. This procedure is very similar to
the methods that check the horizontal and vertical lines.