Add board_full? method
This method checks if the board is full. It uses the board_empty?
method to return false when the board is empty.

Then it uses a loop to obtain and check each column using the
obtain_column and column_full? methods. If one of the columns
is not full, then the method returns false.

When the loop ends, we assume that the board is full, so we return
true.