Add game_over method and helper methods
Added the helper methods announce_winner, that receives a token
and puts a message on the terminal of the winner of the game.

The other helper method is announce_draw. This method just puts
a statement of a draw and returns true.

The game_over method has two parameters: player and coordinates. The
player parameter is the last player that placed a token and we are
verifying if this player made a line (horizontal, vertical
or diagonal). If the player did one of these lines, it calls the
announce_winner method and returns true. If not, it checks if
the board is full. If it's full, it calls the announce_draw method
and returns true. If not it returns false, meaning the game is not
over yet.