Add game method and made changes to some methods
Added the new game method. This methods uses the rest of functions
to play a connect four game. First, it assigns the first player to
a variable. Then it enters an infinite loop. In this loop first it
calls the new turn_message method to indicate the player's turn.
Then it calls the player_turn method. After that, it invokes the
draw_board method to illustrate the player's move. Then it checks
if the current player made a line or if the board is full with
the game_over? method. When one of these conditions applies, it
breaks the loop. Finally, when this method returns false, it
changes the player's variable to the next player and continues the
loop from the beginning.

Another new method is turn_mesasge. It has one parameter: token. This
method just prints the current's player turn.

In the draw_board method, there are now two puts to separate the output
in the terminal and give a better look.

In the announce_winner method, it removed one space to give better look
to the winner message.

Last, the player_turn method now always returns a list with the
coordinates of the token that placed the player. This change is made
to be able to use the game_over method on the connect_four class and
the game method.